requirements.txt
torch
torchvision
numpy
matplotlib
seaborn
scikit-learn
tqdm
Pillow
neptune-client
optuna
plotly
kaleido
timm
